32bitでもビルドできるように再修正

+++ mecab-0.996/src/string_buffer.h
@@ -45,7 +45,7 @@ class StringBuffer {
   StringBuffer& operator<<(unsigned int n)       { _UITOA(n); }
   StringBuffer& operator<<(unsigned long int n)  { _UITOA(n); }
   // modifiered by nanamedou -- start
-  StringBuffer& operator<<(size_t n)  { _UITOA(n); }
+  StringBuffer& operator<<(unsigned long long n)  { _UITOA(n); }
   // modifiered by nanamedou -- end
 #ifdef HAVE_UNSIGNED_LONG_LONG_INT
   StringBuffer& operator<<(unsigned long long int n) { _UITOA(n); }
